requirements.txt
asn1crypto==0.24.0
azure-common==1.1.12
azure-core==1.20.0
azure-identity==1.8.0
azure-keyvault-secrets==4.4.0
azure-mgmt-keyvault==8.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-resource==15.0.0
azure-nspkg==2.0.0
certifi==2018.4.16
cffi==1.15.0
chardet==3.0.4
entrypoints==0.2.3
idna==2.7
isodate==0.6.0
keyring==13.1.0
msrest==0.6.21
msrestazure==0.4.33
oauthlib==2.1.0
pycparser==2.18
PyJWT==1.6.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.12.0
urllib3==1.23
